chore: remove service file removal to filesystem utils
refactor: split filesystem utils into separate files
