+++ cmd/utils/global_flags.go
@@ -1,23 +1,18 @@
 package utils
 
 import (
-	"os"
-	"path/filepath"
-
 	"github.com/spf13/cobra"
+
+	"github.com/dymensionxyz/roller/utils/roller"
 )
 
 func AddGlobalFlags(command *cobra.Command) {
 	command.PersistentFlags().StringP(
-		FlagNames.Home, "", GetRollerRootDir(), "The directory of the roller config files")
+		FlagNames.Home, "", roller.GetRootDir(), "The directory of the roller config files")
 }
 
 var FlagNames = struct {
 	Home string
 }{
 	Home: "home",
-}
-
-func GetRollerRootDir() string {
-	return filepath.Join(os.Getenv("HOME"), ".roller")
 }
