+++ utils/service_manager/service.go
@@ -5,17 +5,12 @@ import (
 	"fmt"
 	"log"
 	"os/exec"
-	"path/filepath"
-	"runtime"
 	"sync"
 	"time"
-
-	"github.com/pterm/pterm"
 
 	"github.com/dymensionxyz/roller/cmd/utils"
 	"github.com/dymensionxyz/roller/utils/bash"
 	"github.com/dymensionxyz/roller/utils/config"
-	"github.com/dymensionxyz/roller/utils/filesystem"
 )
 
 type ServiceConfig struct {
@@ -198,36 +193,6 @@ func StopLaunchdService(serviceName string) error {
 	err := bash.ExecCmd(cmd)
 	if err != nil {
 		return err
-	}
-
-	return nil
-}
-
-func RemoveServiceFiles(services []string) error {
-	if runtime.GOOS == "linux" {
-		pterm.Info.Println("removing old systemd services")
-		for _, svc := range services {
-			svcFileName := fmt.Sprintf("%s.service", svc)
-			svcFilePath := filepath.Join("/etc/systemd/system/", svcFileName)
-
-			err := filesystem.RemoveFileIfExists(svcFilePath)
-			if err != nil {
-				pterm.Error.Println("failed to remove systemd service: ", err)
-				return err
-			}
-		}
-	} else if runtime.GOOS == "darwin" {
-		pterm.Info.Println("removing old systemd services")
-		for _, svc := range services {
-			svcFileName := fmt.Sprintf("xyz.dymension.roller.%s.plist", svc)
-			svcFilePath := filepath.Join("/Library/LaunchDaemons/", svcFileName)
-
-			err := filesystem.RemoveFileIfExists(svcFilePath)
-			if err != nil {
-				pterm.Error.Println("failed to remove systemd service: ", err)
-				return err
-			}
-		}
 	}
 
 	return nil
